Re-organize e2e tests
- Use pytest fixtures for ASG and LaunchTemplate test resources
- Use k8s.wait_on_condition to validate resource synced state
- Split cluster and service tests into their own test_*.py files
